using VoteState enum everywhere
